Rename block modules with correct suffix
Fix import in SiteFooter.
Fix bug in SiteFooter story.
