+++ apps/store/src/blocks/ProductSummaryBlock.tsx
