+++ apps/store/src/components/SiteFooter/SiteFooter.tsx
@@ -2,9 +2,9 @@ import styled from '@emotion/styled'
 import { ChangeEvent, useRef } from 'react'
 import { Space } from 'ui'
 import { InputSelect } from '@/components/InputSelect/InputSelect'
+import { SpaceFlex } from '@/components/SpaceFlex/SpaceFlex'
 import { Locale } from '@/lib/l10n/types'
 import { useCurrentLocale } from '@/lib/l10n/useCurrentLocale'
-import { SpaceFlex } from '../SpaceFlex/SpaceFlex'
 import { Field, MARKET_MAP, TEMP_TRANSLATIONS } from './SiteFooter.constants'
 import { findLocale } from './SiteFooter.helpers'
